requirements.txt
langchain
langchain-groq
langchain-community
wikipedia
duckduckgo-search
python-dotenv
